% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eumap.R
\name{eumap}
\alias{eumap}
\title{Provide a map layer for maps of Europe}
\usage{
eumap(.data, .lon = "lon", .lat = "lat", xlim = c(-10, 30), ylim = c(35,
  60))
}
\arguments{
\item{.data}{data set to plot on the map}

\item{.lon}{name of the longitude variable as character}

\item{.lat}{name of the latitude variable as character}

\item{xlim}{longitudinal boundaries of the map}

\item{ylim}{latitudinal boundaries of the map}
}
\value{
an object of class "gg"
}
