% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmap.R
\name{regmap}
\alias{regmap}
\title{Provide a map layer for maps of various regions}
\usage{
regmap(.data, region = "europe", .lon = "lon", .lat = "lat",
  .xlim = NULL, .ylim = NULL, .region_colour = "#BBBBBB",
  .region_fill = "#DDDDDD", plot = TRUE)
}
\arguments{
\item{.data}{data set to plot on the map}

\item{.lon}{name of the longitude variable as character}

\item{.lat}{name of the latitude variable as character}

\item{.xlim}{longitudinal boundaries of the map}

\item{.ylim}{latitudinal boundaries of the map}

\item{.region_colour}{line colour for the region map}

\item{.region_fill}{fill colour for the region map}

\item{plot}{shall a plot be generated (TRUE, default) or only the
map layer be returned (FALSE)}
}
\value{
an object of class "gg"
}
